include custom exception for missing issue
